+++ java-client/session/build.gradle
@@ -18,7 +18,8 @@ dependencies {
   api project(':proto:proto-backplane-grpc')
 
   api project(':ssl-config')
-  implementation project(':ssl-config-netty')
+  implementation project(':ssl-kickstart')
+  Classpaths.inheritSSLContext(project, 'sslcontext-kickstart-for-netty')
 
   Classpaths.inheritImmutables(project)
   Classpaths.inheritAutoService(project)
