--- server/config-jetty/src/main/java/io/deephaven/ssl/config/package-info.java
@@ -1,7 +1,0 @@
-/**
- * Provides configuration adapters into {@link org.eclipse.jetty.util.ssl.SslContextFactory}.
- *
- * @see io.deephaven.ssl.config.JettySSLConfig#forServer(io.deephaven.ssl.config.SSLConfig)
- * @see io.deephaven.ssl.config.JettySSLConfig#forClient(io.deephaven.ssl.config.SSLConfig)
- */
-package io.deephaven.ssl.config;
