--- server/config-netty/build.gradle
@@ -1,17 +1,0 @@
-plugins {
-    id 'java-library'
-    id 'io.deephaven.project.register'
-}
-
-dependencies {
-    api project(':ssl-config')
-
-    Classpaths.inheritGrpcPlatform(project)
-    compileOnly('io.grpc:grpc-netty') {
-        because 'we want to transitively inherit io.netty:netty-handler'
-    }
-    // We really want a compileOnly dependency on netty-handler, but we can lean on grpc-netty.
-    // compileOnly 'io.netty:netty-handler'
-
-    Classpaths.inheritSSLContext(project, 'sslcontext-kickstart-for-netty')
-}
