--- server/config-netty/src/main/java/io/deephaven/ssl/config/package-info.java
@@ -1,7 +1,0 @@
-/**
- * Provides configuration adapters into {@link io.netty.handler.ssl.SslContextBuilder}.
- *
- * @see io.deephaven.ssl.config.NettySSLConfig#forServer(io.deephaven.ssl.config.SSLConfig)
- * @see io.deephaven.ssl.config.NettySSLConfig#forClient(io.deephaven.ssl.config.SSLConfig)
- */
-package io.deephaven.ssl.config;
