--- server/config/build.gradle
@@ -1,31 +1,0 @@
-plugins {
-    id 'java-library'
-    id 'io.deephaven.project.register'
-}
-
-description = 'The Deephaven SSL Configuration'
-
-dependencies {
-    Classpaths.inheritImmutables(project)
-    Classpaths.inheritJacksonPlatform(project, 'api')
-    Classpaths.inheritGrpcPlatform(project)
-    Classpaths.inheritSSLContext(project, 'sslcontext-kickstart')
-    api 'com.fasterxml.jackson.core:jackson-annotations'
-    implementation 'com.fasterxml.jackson.core:jackson-databind'
-    runtimeOnly('com.fasterxml.jackson.datatype:jackson-datatype-jdk8') {
-        because 'we need to parse optional types'
-    }
-    compileOnly('io.grpc:grpc-core') {
-        because 'we need access to CertificateUtils'
-    }
-
-    Classpaths.inheritJUnitPlatform(project)
-    Classpaths.inheritAssertJ(project)
-
-    testImplementation 'org.junit.jupiter:junit-jupiter'
-    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine'
-}
-
-test {
-    useJUnitPlatform()
-}
