--- server/config/gradle.properties
@@ -1,2 +1,0 @@
-io.deephaven.project.ProjectType=JAVA_PUBLIC
-languageLevel=8
