--- server/config/src/test/resources/io/deephaven/ssl/config/default-explicit.json
@@ -1,18 +1,0 @@
-{
-  "identity": [],
-  "identityProperties": false,
-
-  "trust": [],
-  "trustJdk": true,
-  "trustSystem": false,
-  "trustProperties": false,
-  "trustAll": false,
-
-  "ciphers": [],
-  "ciphersProperties": false,
-
-  "protocols": [],
-  "protocolsProperties": false,
-
-  "clientAuthentication": "NONE"
-}
