+++ ssl/config/build.gradle
@@ -1,0 +1,27 @@
+plugins {
+    id 'java-library'
+    id 'io.deephaven.project.register'
+}
+
+description = 'The Deephaven SSL Configuration'
+
+dependencies {
+    Classpaths.inheritImmutables(project)
+    Classpaths.inheritJacksonPlatform(project, 'api')
+    Classpaths.inheritGrpcPlatform(project)
+    api 'com.fasterxml.jackson.core:jackson-annotations'
+    implementation 'com.fasterxml.jackson.core:jackson-databind'
+    runtimeOnly('com.fasterxml.jackson.datatype:jackson-datatype-jdk8') {
+        because 'we need to parse optional types'
+    }
+
+    Classpaths.inheritJUnitPlatform(project)
+    Classpaths.inheritAssertJ(project)
+
+    testImplementation 'org.junit.jupiter:junit-jupiter'
+    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine'
+}
+
+test {
+    useJUnitPlatform()
+}
