+++ ssl/config/src/test/resources/io/deephaven/ssl/config/default-explicit.json
@@ -1,0 +1,18 @@
+{
+  "identity": [],
+  "identityProperties": false,
+
+  "trust": [],
+  "trustJdk": true,
+  "trustSystem": false,
+  "trustProperties": false,
+  "trustAll": false,
+
+  "ciphers": [],
+  "ciphersProperties": false,
+
+  "protocols": ["TLSv1.2", "TLSv1.3"],
+  "protocolsProperties": false,
+
+  "clientAuthentication": "NONE"
+}
