+++ ssl/config/src/test/resources/io/deephaven/ssl/config/identity-keystore.json
