+++ ssl/config/src/test/resources/io/deephaven/ssl/config/protocols-properties.json
